refactor(models): separate Hybrid MLA provider
